Add country list that displays countries visited
I added a country list which displays the countries that
were visited. I made sure it will not display duplicate
countries by using a reducer.

I also added a Loader component for when the data has not
loaded yet and also added a Message component for when there are no cities in the list yet.