Create reusable Button component
I added a Button component that can be reused in different places and has custome classes for each type of button.

I also began adding URL parameters to create dynamic routes. I have set a query string to the latitude and longitude coordinates which will be useful for when I implement the map.