Create CitiesContext and finish City view.
Created a context to elimate prop drilling and create global state.

I also created the getCity function which gets called evertime the city ID within the URL changes. For this I used a useEffect. It compares the IDs of the cities and then sets the currentCity to the one selected. This means every time we switch to a new city and the ID changes, it will automatically set the new currentCity. Then the currentCity information is displayed in the view.

I also created a reusable BackButton since there is now more than one place where I need it.