Make map change center once city is selected
Once a user clicks on a city from the list, the map will pan and center to that selected city. Also, once the user exits back into the list, the previously selected city will stay there until the next one is selected. I used a useEffect hook to synchronise the map position. Therefore every time the City ID changes in the URL, the lat and lng variables will be updated and the useEffect will set the mapPosition to the new lat and lng coordinates.