Create SetCoords components
This component uses the useMapEvents hook from react-leaflet. After a click on the map happens, the component navigates to the Form component, reads the latlng property from the event object, centers the map to those coordinates and then sets the URL to those coordinates which I will use later in other parts of the application.